made code more modular + componentized
